requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==6.0.1
django-extensions==4.1
django-silk==5.4.3
djangorestframework==3.16.1
gprof2dot==2025.4.14
pillow==12.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
